05.04.217

+++ TranslateIT2v4/src/main/java/translateit2/fileloader/storage/FileSystemStorageService.java
@@ -7,13 +7,17 @@ import org.springframework.stereotype.Service;
 import org.springframework.util.FileSystemUtils;
 import org.springframework.web.multipart.MultipartFile;
 
+import translateit2.util.ISO8859Loader;
 import translateit2.util.OrderedProperties;
 
+import java.io.BufferedReader;
 import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import java.io.InputStreamReader;
+import java.io.UnsupportedEncodingException;
 import java.net.MalformedURLException;
 import java.nio.file.Files;
 import java.nio.file.Path;
@@ -56,20 +60,42 @@ public class FileSystemStorageService implements StorageService {
                 throw new StorageException("Upload directory " + absTgtParent.toString() + 
                 		" was missing and could not be recreated");
             }
-            File t = new File(target.toString());
-            boolean success = Files.deleteIfExists(t.toPath());
-            if (!success) {
-                throw new StorageException("File " + target.toString() + 
+            
+            if (Files.exists(target)) {
+                File t = new File(target.toString());
+                boolean success = Files.deleteIfExists(t.toPath());
+                if (!success) {
+                	throw new StorageException("File " + target.toString() + 
                 		" existed already and could not be removed");
+                }
             }
             /*
              */
-            Files.copy(file.getInputStream(), this.rootLocation.resolve(file.getOriginalFilename()));
+            Path outFilePath = this.rootLocation.resolve(file.getOriginalFilename());           
+            //Files.copy(file.getInputStream(), outFilePath);
+                        
+        	String tmpFilenameStr="temp.properties"; 
+        	Path dir = outFilePath.getParent();        
+            Path fn = outFilePath.getFileSystem().getPath(tmpFilenameStr);
+            Path tmpFilePath = (dir == null) ? fn : dir.resolve(fn);
+            
+            if (Files.exists(tmpFilePath)) {
+                File t = new File(tmpFilePath.toString());
+                boolean success = Files.deleteIfExists(t.toPath());
+                if (!success) {
+                	throw new StorageException("File " + tmpFilePath.toString() + 
+                		" existed already and could not be removed");
+                }
+            }
+            
+            Files.copy(file.getInputStream(), tmpFilePath);            
+            ISO8859Loader.copyISO8859toUTF8(tmpFilePath.toString(), outFilePath.toString());
+            
         } catch (IOException e) {
             throw new StorageException("Failed to store file " + file.getOriginalFilename(), e);
         }
     }
-
+    
     @Override
     public Stream<Path> loadAll() {
         try {
